langchain_arch/requirements.txt
langchain>=0.1.0,<0.2.0
langchain-openai>=0.1.0,<0.2.0
neo4j>=5.0.0,<6.0.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-community>=0.0.10
chainlit>=1.0.0
